requirements.txt
fastapi==0.63.0
uvicorn==0.13.4
asyncpg==0.22.0
pydantic==1.10.2
tortoise-orm==0.19.2
pytest==7.1.3
requests==2.28.1